Fix partialy error on integrity check
By creating all theses files inside the owncloud folder we obtain a lot of failure. In addition by modifying the fiel lib/base.php we create an other failure. The easiest way (but not the proper) is to disable the integrity control by modifyng the value of $OC_Channel in version.php to sth other than 'stable'...
PS : It will be easiest to contribute if you open the issues system on your app, in any case thks for these great job